+++ R/check_deprecated.R
@@ -5,6 +5,8 @@
 #' @param pkg Package that the function is from.
 #' @param args Argument calls to assess.
 #' @param lifecycle_fun Which \pkg{lifecycle} function to use by default.
+#' @param reassign Attempt to reassign deprecated variables to
+#' the corresponding new variable (if applicable).
 #' @param map Mapping between old:new argument names. Use \code{NULL}
 #' if the argument is no longer used at all.
 #' @inheritParams lifecycle::deprecate_warn
@@ -12,6 +14,7 @@
 #' @export
 #' @importFrom stats na.omit
 #' @importFrom lifecycle deprecate_stop
+#' @importFrom utils getFromNamespace
 #' @examples
 #' topSNPs <- echodata::topSNPs_Nalls2019
 #' fullSS_path <- echodata::example_fullSS()
@@ -28,6 +31,7 @@ check_deprecated <- function(fun="finemap_loci",
                              when="2.0.0",
                              args=match.call(),
                              lifecycle_fun=lifecycle::deprecate_warn,
+                             reassign=FALSE,
                              map=list(A1_col="colmap",
                                      A2_col="colmap",
                                      chrom_col="colmap",
@@ -50,7 +54,7 @@ check_deprecated <- function(fun="finemap_loci",
                                      probe_path=NULL,
                                      chrom_type=NULL,
                                      PAINTOR_QTL_datasets=NULL,
-                                     QTL_prefixes=NULL,
+                                     QTL_prefixes="qtl_suffixes",
                                      proportion_cases=NULL,
                                      server=NULL,
                                      vcf_folder=NULL,
@@ -77,8 +81,9 @@ check_deprecated <- function(fun="finemap_loci",
   # args <- fun(4 * atan(1), u = pi)
   # names(args)[-1] <- c("PAINTOR_QTL_datasets","A2_col")#names(map)[seq_len(2)]
 
+
   #### Find all deprecated args ####
-  dep_args <- formals(echolocatoR::finemap_loci)
+  dep_args <- formals(utils::getFromNamespace(fun,pkg))
   dep_args <- sort(
     as.character(
       stats::na.omit(
@@ -93,7 +98,9 @@ check_deprecated <- function(fun="finemap_loci",
   args <- as.list(args)[-1]
   args <- args[names(args) %in% dep_args]
   if(length(args)>0){
-    warning(paste("Found",length(args),"deprecated arguments in use."))
+    wng <- paste("Found",length(args),"deprecated arguments used in:",
+                 deparse(args[[1]]))
+    warning(wng)
     for(nm in names(args)){
       if (lifecycle::is_present(args[[nm]])) {
         ##### Check with ####
@@ -104,20 +111,27 @@ check_deprecated <- function(fun="finemap_loci",
         }
         #### Run appropriate warning/error ####
         if(is.null(map[[nm]])){
-          lifecycle_fun(when = paste0(pkg," ",when),
+          lifecycle_fun(when = when,
                         what = paste0(fun,"(",nm,")"))
         }else if(map[[nm]]=="colmap"){
-            lifecycle::deprecate_stop(when = paste0(pkg," ",when),
+            lifecycle::deprecate_stop(when = when,
                                       what = paste0(fun,"(",nm,")"),
                                       with = with)
 
         }else{
-          lifecycle_fun(when = paste0(pkg," ",when),
+          lifecycle_fun(when = when,
                         what = paste0(fun,"(",nm,")"),
                         with = with)
-          assign(x = map[[nm]],
-                 value = nm,
-                 pos = 3L)
+          #### Reassign values to variables ####
+          if(isTRUE(reassign)){
+            messager("Reassigning deprecated argument:",nm,"-->",map[[nm]])
+            assign(x = map[[nm]],
+                   value = get(nm),
+                   inherits = TRUE,
+                   pos = 1L)
+          } else {
+            messager("Deprecated argument will be ignored:",nm)
+          }
         }
       }
     }