+++ R/imports.R
@@ -1,3 +1,31 @@
+##' catalogueR
+##'
+##' catalogueR
+##' @import catalogueR
+##' @name catalogueR
+NULL
+
+##' downloadR
+##'
+##' downloadR
+##' @import downloadR
+##' @name downloadR
+NULL
+
+##' echoannot
+##'
+##' echoannot
+##' @import echoannot
+##' @name echoannot
+NULL
+
+##' echoconda
+##'
+##' echoconda
+##' @import echoconda
+##' @name echoconda
+NULL
+
 ##' echodata
 ##'
 ##' echodata